combine pointer and gesture input
